baselines/ppo/requirements.txt
absl-py==1.1.0
astor==0.8.1
cached-property==1.5.2
certifi==2022.5.18.1
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
distlib==0.3.4
enum34==1.1.10
filelock==3.4.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.1.0
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.1
munch==2.5.0
networkx==2.5.1
numpy==1.18.5
opencv-python==4.4.0.42
opt-einsum==3.3.0
pathlib==1.0.1
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.4
psutil==5.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
six==1.16.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.4
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tqdm==4.56.0
typing-extensions==4.1.1
virtualenv==20.13.2
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
